Make cuse-dev-create a macro to accept varargs
In order to accept the varargs parameters to the format string we have
to supply the correct types to foreign-procedure.  Make this somewhat
ergonomic by defining cuse-dev-create as a macro that takes the argument
types to append to the binding.  It would be even nicer to infer this
from the format string, someday.